pm/hydra: Store proxy count in process group structure
Not having access to the number of proxies in a process group meant we
had to count them by traversing the list at certain times. This made for
an inefficient barrier algorithm that counted the number of proxies each
time it received a proxy command. Instead, we store the number of
proxies in the pg struct so we can compare it against the current
barrier count.